refactor: move request body to a Payload.hs module
